完成part4-d的作业部分

+++ part4/exercise-bloglist/controllers/blogs.js
@@ -1,14 +1,21 @@
+const jwt = require('jsonwebtoken')
 const express = require('express')
 const blogRouter = express.Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogRouter.post('/', async (request, response) => {
   const blog = new Blog(request.body)
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
 
   if (!blog.title || !blog.url) {
     response.status(400).send('Bad Request')
@@ -18,10 +25,23 @@ blogRouter.post('/', async (request, response) => {
   blog.likes = blog.likes || 0
 
   const savedBlog = await blog.save()
+  const user = await User.findById(blog.user)
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(200).json(savedBlog)
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  const userId = decodedToken.id
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== userId) {
+    return response.status(401).send({ error: 'token missing or invalid' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
